Implemented delay timer and related opcodes
time.c has some experimenting I did.  I wanted to have the timer run
in "real time" and not rely on a sleep function which would halt
execution.  That way it could run multiple instructions per tick
of the timer.  Had some difficulty with this because the clock runs at 60Hz
and most of the functions to get a time aren't that precise.  Ended up using usleep for now