Added a function to look up opcodes
This function takes the instruction and returns a number specifying
the opcode.  This number can then be used in other switch statements
or as array indices so it is simpler to map functions to specific
opcodes.  This can be used in both a stepping function and a
disassembler, to start with.