Created DecodedInstr type for disassembler
The DecodedInstr type is just an Instruction with some additional
data associated with it.  This includes things like whether it is
a special jump instruction (along with other flags TBD) and a
buffer with the human readable format (as printed by dasm).  The
new type isn't fully integrated or utilized yet but that is the
next step.